requirements.txt
fastapi
uvicorn
pydantic
playwright
google-generativeai
python-dotenv
firebase-admin